# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "title": "Pool API documentation",
        "contact": {},
        "version": "3.0.0"
    },
    "host": "pool3-api.vivaconagua.org",
    "basePath": "/v1",
    "paths": {
...
        "/events/event": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/events/event"
                ],
                "summary": "Get a List of Event",
                "parameters": [
                    {
                        "type": "string",
                        "name": "createdFrom",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "createdTo",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "name": "crewID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "endAt",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "eventASPID",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "name": "eventState",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "fullCount",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "internalASPID",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "missingApplications",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "onlyApply",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "name": "organisationId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "sortDirection",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "sortField",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "startAt",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "updatedFrom",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "updatedTo",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/vcago.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Event"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/vcago.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/events/event"
                ],
                "summary": "Update a Event",
                "parameters": [
                    {
                        "description": "Event Data",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.EventUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/vcago.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "$ref": "#/definitions/models.Event"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "creates an Event object.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/events/event"
                ],
                "summary": "Create a Event",
                "parameters": [
                    {
                        "description": "Event Data",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.EventCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/vcago.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "$ref": "#/definitions/models.Event"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/events/event/details/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/events/event"
                ],
                "summary": "Get a Event by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/vcago.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "$ref": "#/definitions/models.EventDetails"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/events/event/email": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/events/event"
                ],
                "summary": "Get a List of Event",
                "parameters": [
                    {
                        "type": "string",
                        "name": "createdFrom",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "createdTo",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "name": "crewID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "endAt",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "eventASPID",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "name": "eventState",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "fullCount",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "name": "id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "internalASPID",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "missingApplications",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "name": "onlyApply",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "name": "organisationId",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "sortDirection",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "sortField",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "startAt",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "csv",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "updatedFrom",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "name": "updatedTo",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/vcago.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.EventPublic"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/vcago.Response"
                        }
                    }
                }
            }
        },
...
        "/events/event/sync/{id}": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/events/event"
                ],
                "summary": "Sync Event by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/vcago.Response"
                        }
                    }
                }
            }
        },
...
        "/events/event/view/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/events/event"
                ],
                "summary": "Get a Event by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/vcago.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "$ref": "#/definitions/models.EventPublic"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/events/event/{id}": {
            "get": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/events/event"
                ],
                "summary": "Get a Event by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/vcago.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "payload": {
                                            "$ref": "#/definitions/models.Event"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "CookieAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "/events/event"
                ],
                "summary": "Delete a Event by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/vmod.DeletedResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.EventDetails": {
            "type": "object",
            "properties": {
                "event_asp": {
                    "$ref": "#/definitions/models.EventASPPublic"
                },
                "meeting_url": {
                    "type": "string"
                }
            }
        },
...
        "models.EventUpdate": {
            "type": "object",
            "properties": {
                "additional_information": {
                    "type": "string"
                },
                "application": {
                    "$ref": "#/definitions/models.EventApplication"
                },
                "artist_ids": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "crew": {
                    "$ref": "#/definitions/models.Crew"
                },
                "crew_id": {
                    "type": "string"
                },
                "end_at": {
                    "type": "integer"
                },
                "event_asp_id": {
                    "type": "string"
                },
                "event_state": {
                    "$ref": "#/definitions/models.EventState"
                },
                "event_tools": {
                    "$ref": "#/definitions/models.EventTools"
                },
                "external_asp": {
                    "$ref": "#/definitions/models.UserExternal"
                },
                "id": {
                    "type": "string"
                },
                "internal_asp_id": {
                    "type": "string"
                },
                "location": {
                    "$ref": "#/definitions/models.Location"
                },
                "meeting_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "organisation_id": {
                    "type": "string"
                },
                "organizer_id": {
                    "type": "string"
                },
                "start_at": {
                    "type": "integer"
                },
                "tour_id": {
                    "type": "string"
                },
                "type_of_event": {
                    "type": "string"
                },
                "website": {
                    "type": "string"
                }
            }
        },
...
        "vcago.Response": {
            "type": "object",
            "properties": {
                "list_size": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "model": {
                    "type": "string"
                },
                "payload": {},
                "type": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Bearer key",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CookieAuth": {
            "type": "apiKey",
            "name": "access_cookie",
            "in": "cookie"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://hydra.vivaconagua.org/oauth/authorize",
            "tokenUrl": "https://hydra.vivaconagua.org/oauth/token"
        }
    }
}